server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
numpy==1.26.4
scipy==1.14.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
motor==3.6.0
bcrypt==4.2.0
PyJWT==2.9.0
